fix: 修复类型交叉引用的问题
