+++ packages/field/src/index.tsx
@@ -41,6 +41,7 @@ import FieldImage from './components/Image';
 import FieldCascader from './components/Cascader';
 import FieldColorPicker from './components/ColorPicker';
 import FieldDigitRange from './components/DigitRange';
+// import type {RangeInputNumberProps,ExtraProps as } from './components/DigitRange'
 import { noteOnce } from 'rc-util/lib/warning';
 
 const REQUEST_VALUE_TYPE = ['select', 'radio', 'radioButton', 'checkbook'];
